pyproject.toml
[project]
name = "astra"
version = "0.1.0"
description = "Astra is an AI Backend-as-a-Service platform. It provides a high-performance runtime for building, executing, and scaling AI agents with memory, orchestration, and observability."
readme = "README.md"
requires-python = ">=3.10,<3.14"
authors = [
    { name = "Himanshu Sharma", email = "himanshu.kumarr07@gmail.com" }
]
dependencies = [
    "aiohttp>=3.13.2",
    "boto3>=1.40.61",
    "debugpy>=1.8.17",
    "opentelemetry-sdk>=1.38.0",
]


[tool.uv.workspace]
members = [
    # "packages/deployment",  # Empty package, commented out temporarily
    "packages/framework",
    # "packages/integrations",  # Empty package, commented out temporarily
    "packages/memorybase",
    "packages/observability",
    # "packages/runtime",  # Empty package, commented out temporarily
]

[tool.uv.sources]
observability = { workspace = true }
framework = { workspace = true }
memorybase = { workspace = true }
# runtime = { workspace = true }  # Empty package

[tool.ruff]
target-version = "py310"
line-length = 100
respect-gitignore = true
force-exclude = true
fix = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
future-annotations = true
extend-select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes
    "B",      # flake8-bugbear
    "RUF",    # ruff-specific rules
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "I",      # isort
    "T",      # flake8-print
    "PERF",   # performance issues
    "PIE",    # flake8-pie
    "TID",    # flake8-tidy-imports
    "ASYNC",  # async/await best practices
    "UP006",  # use collections.abc
    "UP007",  # use X | Y for unions
    "UP035",  # use dict/list instead of typing.Dict/List
    "UP037",  # remove quotes from type annotations
    "UP045",  # use X | None instead of Optional[X]
    "UP004",  # use isinstance instead of type
    "UP008",  # use super() instead of super(Class, self)
    "UP010",  # use isinstance for type checks
    "UP018",  # use str() instead of "string"
    "UP031",  # use f-strings for .format()
    "UP032",  # use f-strings for .format() with positional
    "I001",   # sort imports
    "I002",   # remove unused imports
]
ignore = ["E501"]  # ignore line too long globally

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
no-sections = false
case-sensitive = true
combine-as-imports = true
force-single-line = false
force-sort-within-sections = true
known-first-party = []
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-after-imports = 2
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"**/examples/**" = ["T201"]  # allow print in examples

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
show_error_codes = true

[dependency-groups]
dev = [
    "framework",
    "memorybase",
    "observability",
    "ruff>=0.14.6",
]
